third_party/blink/web_tests/wpt_internal/shared_storage/run-url-selection-operation.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<script src="/common/utils.js"></script>
<script src="../fenced_frame/resources/utils.js"></script>

<body>
<script>
'use strict';

promise_test(async () => {
  const ancestor_key = token();
  let url0 = generateURL("resources/frame0.html", [ancestor_key]);
  let url1 = generateURL("resources/frame1.html", [ancestor_key]);

  await sharedStorage.worklet.addModule("resources/simple-module.js");

  let uuid0 = await sharedStorage.selectURL(
      "test-url-selection-operation", [{url: url0}, {url: url1}],
      {data: {'mockResult': 0}});
  assert_true(uuid0.startsWith('urn:uuid:'));
  attachFencedFrame(uuid0, 'opaque-ads');
  const result0 = await nextValueFromServer(ancestor_key);
  assert_equals(result0, "frame0_loaded");

  let uuid1 = await sharedStorage.selectURL(
      "test-url-selection-operation", [{url: url0}, {url: url1}],
      {data: {'mockResult': 1}});
  assert_true(uuid1.startsWith('urn:uuid:'));
  attachFencedFrame(uuid1, 'opaque-ads');
  const result1 = await nextValueFromServer(ancestor_key);
  assert_equals(result1, "frame1_loaded");

  let uuid2 = await sharedStorage.selectURL(
      "test-url-selection-operation", [{url: url0}, {url: url1}],
      {data: {'mockResult': -1}});
  assert_true(uuid2.startsWith('urn:uuid:'));
  attachFencedFrame(uuid2, 'opaque-ads');
  const result2 = await nextValueFromServer(ancestor_key);
  assert_equals(result2, "frame0_loaded");
}, 'selectURL()');

</script>
</body>
